Renamed CosmicRegistries.java to CosmicRegistration.java
