+++ src/main/java/com/ghostipedia/cosmiccore/api/registries/CosmicRegistration.java
@@ -2,6 +2,6 @@ package com.ghostipedia.cosmiccore.api.registries;
 import com.ghostipedia.cosmiccore.CosmicCore;
 import com.gregtechceu.gtceu.api.registry.registrate.GTRegistrate;
 
-public class CosmicRegistries {
+public class CosmicRegistration {
  public static final GTRegistrate REGISTRATE = GTRegistrate.create(CosmicCore.MOD_ID);
 }
