+++ src/main/java/com/ghostipedia/cosmiccore/common/data/materials/CosmicMaterials.java
@@ -56,7 +56,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1240))
                 .color(0xff8400).secondaryColor(0xcffee00).iconSet(MaterialIconSet.SHINY)
-                .flags(GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .components(Virtue, 4, Naquadah, 1,TungstenSteel, 15, Chromium, 3, Molybdenum, 6, Vanadium, 3,Prisma,8)
                 .cableProperties(GTValues.V[5], 4, 6)
                 .blastTemp(4200, BlastProperty.GasTier.HIGHER, GTValues.VA[GTValues.IV], 1800)
@@ -65,7 +65,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1300))
                 .color(0x001f04).secondaryColor(0x000000).iconSet(MaterialIconSet.RADIOACTIVE)
-                .flags(GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .components(BlackSteel, 4, NaquadahEnriched, 3, RhodiumPlatedPalladium, 6)
                 .cableProperties(GTValues.V[6], 4, 6)
                 .blastTemp(5400, BlastProperty.GasTier.HIGHER, GTValues.VA[GTValues.LuV],2400)
@@ -75,7 +75,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1500))
                 .color(0xdb5e5e).secondaryColor(0x5e0f3d).iconSet(MaterialIconSet.BRIGHT)
-                .flags(GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .cableProperties(GTValues.V[7], 4, 6)
                 .blastTemp(7000, BlastProperty.GasTier.HIGHER, GTValues.VA[GTValues.ZPM],5600)
                 .buildAndRegister();
@@ -84,7 +84,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1760))
                 .color(0x940002).secondaryColor(0xe35e17)
-                .flags(GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .cableProperties(GTValues.V[8], 4, 6)
                 .blastTemp(9000, BlastProperty.GasTier.HIGHER, GTValues.VA[GTValues.UV],5600)
                 .buildAndRegister();
@@ -92,7 +92,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1760))
                 .color(0x004094).secondaryColor(0x0962ab)
-                .flags(GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .cableProperties(GTValues.V[9], 4, 6)
                 .blastTemp(10500, BlastProperty.GasTier.HIGHER, GTValues.VA[GTValues.UV],5600)
                 .buildAndRegister();
@@ -100,7 +100,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1760))
                 .color(0x003e4f).secondaryColor(0x0a2830)
-                .flags(GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_ROUND,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .cableProperties(GTValues.V[10], 4, 6)
                 .blastTemp(13000, BlastProperty.GasTier.HIGHEST, GTValues.VA[GTValues.UHV],5600)
                 .buildAndRegister();
@@ -108,7 +108,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1760))
                 .color(0xc7e9f2).secondaryColor(0xffffff)
-                .flags(GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .cableProperties(GTValues.V[11], 4, 6)
                 .blastTemp(15000, BlastProperty.GasTier.HIGHEST, GTValues.VA[GTValues.UEV],5600)
                 .buildAndRegister();
@@ -116,7 +116,7 @@ public class CosmicMaterials {
                 .ingot()
                 .liquid(new FluidBuilder().temperature(1760))
                 .color(0x270054).secondaryColor(0x481c7a)
-                .flags(GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
+                .flags(GENERATE_BOLT_SCREW,GENERATE_GEAR, GENERATE_SMALL_GEAR, GENERATE_RING, GENERATE_FRAME, GENERATE_SPRING, GENERATE_SPRING_SMALL, GENERATE_FINE_WIRE)
                 .cableProperties(GTValues.V[11], 4, 6)
                 .blastTemp(19000, BlastProperty.GasTier.HIGHEST, GTValues.VA[GTValues.UIV],5600)
                 .buildAndRegister();
